refactor: Rename web -> webui and ui -> webui directories
Renamed directories for consistency:
- web/ -> webui/ (frontend source)
- internal/app/coordinator/ui/ -> internal/app/coordinator/webui/ (Go embed)

Updated all references in Makefile, Dockerfile, .gitignore, and server.go.
