+++ internal/app/coordinator/server.go
@@ -16,7 +16,7 @@ import (
 	"github.com/strrl/wonder-mesh-net/internal/app/coordinator/database"
 	"github.com/strrl/wonder-mesh-net/internal/app/coordinator/repository"
 	"github.com/strrl/wonder-mesh-net/internal/app/coordinator/service"
-	"github.com/strrl/wonder-mesh-net/internal/app/coordinator/ui"
+	"github.com/strrl/wonder-mesh-net/internal/app/coordinator/webui"
 	"github.com/strrl/wonder-mesh-net/pkg/apikey"
 	"github.com/strrl/wonder-mesh-net/pkg/headscale"
 	"github.com/strrl/wonder-mesh-net/pkg/jointoken"
@@ -363,7 +363,7 @@ func (s *Server) Run() error {
 	})
 
 	// Web UI - served at /ui/
-	uiHandler, err := ui.Handler()
+	uiHandler, err := webui.Handler()
 	if err != nil {
 		return fmt.Errorf("initialize ui handler: %w", err)
 	}
