+++ internal/app/coordinator/webui/embed.go
@@ -1,4 +1,4 @@
-package ui
+package webui
 
 import (
 	"embed"
