+++ internal/app/coordinator/webui/static/index.html
